added settings, need to chenge branch
